CJH fix for subclassing kinds:
per CJH:
"
I have found a way to subclass kinds in the current context.
To do this we need to be able to override data.get
so I have changed the order of base classes so the app class takes precedence over the data class.  (serve/app.py +158)
We may have needed this to happen for a decade or so...

Now, when the overridden get is called it looks for a class within Page named Page_<kind>

eg.

class Page:
  ...
  class Page_somekind:
     "override methods here"

Maybe this isn't the most intuitive arrangement but at the moment it is the path of least resistance.
In the short term it will be most useful for overriding view and edit_form.

Following through from this I have altered the way Evo template filenames are tried so it will fall back to the un-overridden template if the override isn't found.
"